put inout and outputs to component class. Also decreased the time for dragging to start from 100 ms to  8 ms

+++ src/LogicElements/LogicElementBase.cpp
@@ -443,7 +443,7 @@ namespace LogicElements
             if (c->sourceGate.get() == gate || c->targetGate.get() == gate)
             {
                 std::vector<unsigned int> indices_destroyed;
-                for (size_t j = c->physCon.wires.size() - 1; j > 1; j--)
+                for (int j = c->physCon.wires.size() - 1; j > 1; j--)
                 {
                     Vector2 end = c->physCon.wires[j];
                     Vector2 mid = c->physCon.wires[j - 1];
@@ -459,7 +459,7 @@ namespace LogicElements
                     c->physCon.wires.erase(c->physCon.wires.begin() + indices_destroyed[j]);
                 }
                 indices_destroyed.clear();
-                for (size_t j = c->physCon.wires.size() - 1; j > 1; j--)
+                for (int j = c->physCon.wires.size() - 1; j > 1; j--)
                 {
                     Vector2 end = c->physCon.wires[j];
                     Vector2 mid = c->physCon.wires[j - 1];
